added enemies and hud classes

+++ Dragon Tale Tutorial/src/TileMap/TileMap.java
@@ -105,8 +105,8 @@ public class TileMap
 	}
 	
 	public int getTileSize() { return tileSize; }
-	public int getx() { return (int) x; }
-	public int gety() { return (int) y; }
+	public double getx() { return  x; }
+	public double gety() { return  y; }
 	public int getWidth() { return width; }
 	public int getHeight() { return height; }
